+++ CHANGES.md
@@ -1,10 +1,17 @@
 
+# Changelog
+
+## History
+
+v0.1.2
+
+* Change default behavior to be async.
+* rpc.invoke callback (err, data), data has extra level of data (data.data)
+
 
 ## Future
 
-* Change default behavior to be async.
 * Add error response function `res.error()`
-* rpc.invoke callback (err, data), data has extra level of data (data.data)